Placed all methods in correct Data layers. Changed from Newtonsoft.Json package to the standard dotnet System.Text.Json package
