+++ ProjectTest/TicketSystemTest.cs
@@ -3,9 +3,9 @@ namespace ProjectTest.TicketSystemTest;
 public class TicketSystemTest{
     [TestMethod]
     public void TestMethodJsonUpdate(){
-        Ticket.JsonTestPath = "../../../Sources/tickets.json";
-        // Ticket testTicket = new Ticket("Movie", "01-02-2000", "12:00", "L");
+        UpdateTicketJson.JsonTestPath = "../../../Sources/tickets.json";
+        Ticket testTicket = new Ticket("Movie", "01-02-2000", "12:00", "L");
         Assert.IsTrue(!string.IsNullOrEmpty(File.ReadAllText("../../../Sources/tickets.json")));
-        Ticket.JsonTestPath = null;
+        UpdateTicketJson.JsonTestPath = null;
     }
 }
